Splitted modules aws_clients and aws_roles
