+++ clouds/aws/rhoam_add_role_policy.py
@@ -1,4 +1,5 @@
 import aws_clients
+from aws_roles import AWSRolePolicies
 from simple_logger.logger import get_logger
 
 LOGGER = get_logger(name=__name__)
@@ -10,7 +11,7 @@ JSON_FILE = "rhoam-sre-support-policy.json"
 
 def main():
     iam_client = aws_clients.IAMClient()
-    iam_client.create_role_policy(
+    AWSRolePolicies(client=iam_client).create_role_policy(
         role_name=ROLE_NAME, policy_name=POLICY_NAME, policy_json_path=JSON_FILE
     )
